+++ src/app/constants.ts
@@ -1,2 +1,1 @@
 export const baseUrl = 'https://rickandmortyapi.com/api';
-export const charactersLink = `${baseUrl}/character`;
