+++ src/pages/about-us/AboutUs.tsx
@@ -1,4 +1,4 @@
-import styles from './index.module.scss';
+import styles from './AboutUs.module.scss';
 
 export const AboutUs = () => (
   <div
