+++ src/pages/about-us/index.ts
@@ -1,0 +1,1 @@
+export { AboutUs } from './AboutUs';
