+++ src/pages/home/models.ts
@@ -18,15 +18,3 @@ export interface Character {
   url: string;
   created: string;
 }
-
-export const CardDataKeys = [
-  'id',
-  'name',
-  'imgUrl',
-  'breed',
-  'description',
-] as const;
-
-export type CardData = {
-  [key in (typeof CardDataKeys)[number]]: string;
-};
