+++ src/pages/home/modules/Card/Card.models.ts
