+++ src/pages/home/modules/Card/index.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { vi } from 'vitest';
 
-import { Card } from '.';
+import { Card } from './Card';
 import { MOCK_CHARACTER } from '../../../../app/mocks/data';
 
 const TEST_ID = 'card';
