+++ src/pages/home/modules/CardWithInfo/CardWithInfo.models.ts
