+++ src/pages/home/modules/CardWithInfo/CardWithInfo.module.scss
