+++ src/pages/home/modules/index.ts
@@ -1,4 +1,4 @@
 export { CardList } from './CardList/';
-export { Card } from './Card';
+export { Card } from './Card/Card';
 export { SearchBar } from './SearchBar/SearchBar';
-export { CardWithInfo } from './CardWithInfo';
+export { CardWithInfo } from './CardWithInfo/CardWithInfo';
