+++ src/shared/libs/classNames/index.ts
@@ -2,10 +2,12 @@
 
 export const classNames = (
   cls: string,
-  mods: Record<string, boolean | string>,
+  mods: Record<string, boolean | string> = {},
+  additional: string[] = []
 ): string => {
   return [
     cls,
+    ...additional.filter(Boolean),
     ...Object.entries(mods)
       .filter(([cls, value]) => Boolean(value))
       .map(([cls]) => cls),
