+++ src/widgets/footer/Footer.module.scss
