+++ src/widgets/footer/Footer.tsx
@@ -1,3 +1,3 @@
-import styles from './index.module.scss';
+import styles from './Footer.module.scss';
 
 export const Footer = () => <footer className={styles.footer}>RSS 2023</footer>;
