+++ src/widgets/header/modules/heading/index.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
 
-import Heading from '.';
+import Heading from './heading';
 
 const TEST_ID = 'heading';
 const DEFAULT_TITLE = 'Home';
