+++ src/widgets/header/modules/nav-bar/NavBar.module.scss
