+++ src/widgets/header/modules/nav-bar/constants.ts
