+++ src/widgets/index.ts
@@ -1,2 +1,2 @@
-export { Header } from './header';
-export { Footer } from './footer';
+export { Header } from './header/Header';
+export { Footer } from './footer/Footer';
